backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
openai==0.27.0
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.0.1 